+++ NEWS.md
@@ -1,5 +1,7 @@
 # admtools (development version)
 
+* integration with the `FossilSim` package, specifically transformation of `taxonomy` and `fossils` objects between time and stratigraphic domain
+
 # admtools 0.4.0
 
 * added transformations for S3 class `pre_paleoTS` to integrate with `StratPal` and `paleoTS` package